Add a way to update objects on stripe based on current ones
It's the purpose of the `api_modify` method

It can accept a `fields` arguments, which is an iterable of field names
to take from the current object and pass to stripe. Also, all arguments
passed in `kwargs` are passed to stipe. So it's possible to set fields
that are not in the current object.